Make confidence tests compatible with Cygwin.

+++ Makefile
@@ -34,10 +34,8 @@ lock:
 unlock:
 	sudo yum versionlock clear
 
-test:
-	echo "You don't want to run this."
-	exit 1
-	pkill gpg-agent ; rm -rf /tmp/tmp.*
-	export PATH=~/gitwork/blackbox/bin:/usr/local/bin:/bin:/usr/bin:/usr/local/sbin:/usr/sbin:/sbin ; \
+confidence:
+	@if [[ -e ~/.gnupg ]]; then echo ERROR: '~/.gnupg should not exist. If it does, bugs may polute your .gnupg configuration. If the code has no bugs everything will be fine. Do you feel lucky?'; false ; fi
+	if which >/dev/null gpg-agent ; then pkill gpg-agent ; rm -rf /tmp/tmp.* ; fi
+	export PATH=~/gitwork/blackbox/bin:/usr/local/bin:/bin:/usr/bin:/usr/local/sbin:/usr/sbin:/sbin ; 
 		cd ~/gitwork/blackbox && tools/confidence_test.sh
-	@if [[ -e ~/.gnupg ]]; then echo ERROR: '~/.gnupg' should not exist. If it does, this means test test suite may be poluting your actual .gnupg configuration. ; false ; fi
